implemented VRDevice proximity metadata

+++ src/main/java/com/calicraft/vrjester/gesture/radix/Trace.java
@@ -17,7 +17,7 @@ public class Trace {
     public String voxId; // The Vox ID
     public VRDevice vrDevice; // The VRDevice
     public String movement = "idle"; // Movement taken to get to Vox
-    public final Map<String, Integer> devicesInProximity = new HashMap<>(); // Other VRDevices within this Vox
+    public final Map<String, Long> devicesInProximity = new HashMap<>(); // Other VRDevices within this Vox
     private long elapsedTime = 0; // Time spent within Vox in ms (added on the fly while idle)
     private double speed; // Average speed within Vox (calculated on the fly while idle)
     private Vec3 faceDirection, direction, front, back, right, left,
@@ -107,11 +107,11 @@ public class Trace {
         this.direction = direction;
     }
 
-    public void addDeviceInProximity(String vrDevice, Integer elapsedTime) {
+    public void addDeviceInProximity(String vrDevice, Long elapsedTime) {
         devicesInProximity.put(vrDevice, elapsedTime);
     }
 
-    public Map<String, Integer> getDevicesInProximity() {
+    public Map<String, Long> getDevicesInProximity() {
         return devicesInProximity;
     }
 
@@ -129,6 +129,7 @@ public class Trace {
         setMovement(gestureDirection);
         setElapsedTime(System.nanoTime());
         setSpeed(end);
+        // TODO - Implement way to calculate elapsed time for devices in proximity
     }
 
     private void setMovementBuckets(Vec3 faceDirection) {
